# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Consumer Product Recalls API",
    "description": "Open, read-only API for US consumer product recalls from five agencies (CPSC, FDA, USDA, NHTSA, USCG). No key, GET-only, cursor-paginated. A few things worth knowing up front: a recall's `is_active` flag is `null` (not false) for CPSC and NHTSA, which don't track an open/closed status; `classification` uses each agency's own severity scale and can't be compared across agencies; UPC search matches a whole recall, not an individual product; and search is exact — a typo finds nothing.\n\n**Common lookups:** by product name → `GET /products/search?q=`; by **UPC barcode** → `GET /products/search?upc=`; vehicle or boat by identifier → `GET /products/search?model=` / `?hin=`; a single recall → `GET /recalls/{source}/{recall_id}`.",
    "version": "0.1.1"
  },
  "paths": {
...
    "/stats/overview": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Headline KPIs: total recalls, distinct firms, sources covered, last gold rebuild.",
        "operationId": "overview_stats_overview_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsOverview"
                }
              }
            }
          },
          "400": {
            "description": "BadCursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "InvalidParameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "UpstreamUnavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/stats/recalls-by-period": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Recall counts per period (month/week/year) per source + 'ALL' rollup.",
        "description": "Recall counts per period (month, week, or year) for each agency, plus an `ALL` rollup. Periods are dated by when each recall was first announced, falling back to its publish date when no announcement date is on record — so counts can differ from a publish-date view.",
        "operationId": "recalls_by_period_stats_recalls_by_period_get",
        "parameters": [
          {
            "name": "grain",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Grain",
              "description": "Time grain: month | week | year.",
              "default": "month"
            },
            "description": "Time grain: month | week | year."
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/StatsSource"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one agency, or `ALL` for the all-agency rollup where available. Omit to get every row; an agency with no data for this stat returns an empty list.",
              "title": "Source"
            },
            "description": "Filter to one agency, or `ALL` for the all-agency rollup where available. Omit to get every row; an agency with no data for this stat returns an empty list."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PeriodCount"
                  },
                  "title": "Response Recalls By Period Stats Recalls By Period Get"
                }
              }
            }
          },
          "400": {
            "description": "BadCursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "InvalidParameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "UpstreamUnavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/stats/monthly-trend": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Per-source monthly trend with rolling averages + year-over-year change.",
        "description": "Per-agency monthly counts with rolling averages and year-over-year change. Months are dated by when each recall was first announced (or its publish date when none is on record).",
        "operationId": "monthly_trend_stats_monthly_trend_get",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/StatsSource"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one agency, or `ALL` for the all-agency rollup where available. Omit to get every row; an agency with no data for this stat returns an empty list.",
              "title": "Source"
            },
            "description": "Filter to one agency, or `ALL` for the all-agency rollup where available. Omit to get every row; an agency with no data for this stat returns an empty list."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MonthlyTrendPoint"
                  },
                  "title": "Response Monthly Trend Stats Monthly Trend Get"
                }
              }
            }
          },
          "400": {
            "description": "BadCursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "InvalidParameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "UpstreamUnavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/stats/by-classification": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Recall counts by source-native classification + risk_level (+ 'ALL').",
        "operationId": "by_classification_stats_by_classification_get",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/StatsSource"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one agency, or `ALL` for the all-agency rollup where available. Omit to get every row; an agency with no data for this stat returns an empty list.",
              "title": "Source"
            },
            "description": "Filter to one agency, or `ALL` for the all-agency rollup where available. Omit to get every row; an agency with no data for this stat returns an empty list."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClassificationCount"
                  },
                  "title": "Response By Classification Stats By Classification Get"
                }
              }
            }
          },
          "400": {
            "description": "BadCursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "InvalidParameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "UpstreamUnavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/stats/status": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Active/inactive/unknown recall counts per source (+ 'ALL').",
        "operationId": "status_stats_status_get",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/StatsSource"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one agency, or `ALL` for the all-agency rollup where available. Omit to get every row; an agency with no data for this stat returns an empty list.",
              "title": "Source"
            },
            "description": "Filter to one agency, or `ALL` for the all-agency rollup where available. Omit to get every row; an agency with no data for this stat returns an empty list."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StatusCount"
                  },
                  "title": "Response Status Stats Status Get"
                }
              }
            }
          },
          "400": {
            "description": "BadCursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "InvalidParameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "UpstreamUnavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/stats/firm-leaderboard": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Most-recalled firms, ranked (top-N).",
        "operationId": "firm_leaderboard_stats_firm_leaderboard_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Top-N firms (1-100).",
              "default": 20,
              "title": "Limit"
            },
            "description": "Top-N firms (1-100)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FirmLeaderRow"
                  },
                  "title": "Response Firm Leaderboard Stats Firm Leaderboard Get"
                }
              }
            }
          },
          "400": {
            "description": "BadCursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "InvalidParameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "UpstreamUnavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/stats/by-geography": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Per-state recall counts; basis = distribution | firm_registration (+ 'ALL').",
        "operationId": "by_geography_stats_by_geography_get",
        "parameters": [
          {
            "name": "basis",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/GeographyBasis",
              "description": "distribution | firm_registration (different questions, not a toggle).",
              "default": "distribution"
            },
            "description": "distribution | firm_registration (different questions, not a toggle)."
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/StatsSource"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one agency, or `ALL` for the all-agency rollup where available. Omit to get every row; an agency with no data for this stat returns an empty list.",
              "title": "Source"
            },
            "description": "Filter to one agency, or `ALL` for the all-agency rollup where available. Omit to get every row; an agency with no data for this stat returns an empty list."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GeographyCount"
                  },
                  "title": "Response By Geography Stats By Geography Get"
                }
              }
            }
          },
          "400": {
            "description": "BadCursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "InvalidParameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "UpstreamUnavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/stats/by-country": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Per-distribution-country recall counts (FDA/USDA + 'ALL').",
        "operationId": "by_country_stats_by_country_get",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/StatsSource"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one agency, or `ALL` for the all-agency rollup where available. Omit to get every row; an agency with no data for this stat returns an empty list.",
              "title": "Source"
            },
            "description": "Filter to one agency, or `ALL` for the all-agency rollup where available. Omit to get every row; an agency with no data for this stat returns an empty list."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryCount"
                  },
                  "title": "Response By Country Stats By Country Get"
                }
              }
            }
          },
          "400": {
            "description": "BadCursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "InvalidParameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "UpstreamUnavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/stats/units": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Units recalled per agency and unit type, by month (not comparable across agencies).",
        "description": "Units recalled per agency and unit type, by month. Units come in incomparable kinds (counts, weight, volume), so never add them across kinds or agencies. Months are dated by when each recall was first announced (or its publish date when none is on record).",
        "operationId": "units_stats_units_get",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/StatsSource"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one agency, or `ALL` for the all-agency rollup where available. Omit to get every row; an agency with no data for this stat returns an empty list.",
              "title": "Source"
            },
            "description": "Filter to one agency, or `ALL` for the all-agency rollup where available. Omit to get every row; an agency with no data for this stat returns an empty list."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UnitsRow"
                  },
                  "title": "Response Units Stats Units Get"
                }
              }
            }
          },
          "400": {
            "description": "BadCursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "InvalidParameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "UpstreamUnavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope"
      },
...
      "Grain": {
        "type": "string",
        "enum": [
          "month",
          "week",
          "year"
        ],
        "title": "Grain",
        "description": "Time grain for ``/stats/recalls-by-period``."
      },
...
      "PeriodCount": {
        "properties": {
          "period": {
            "type": "string",
            "format": "date",
            "title": "Period",
            "description": "Period start (month / ISO-week Monday / Jan-1 per grain)."
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Agency feed, or 'ALL' for the all-source rollup."
          },
          "event_count": {
            "type": "integer",
            "title": "Event Count",
            "description": "Distinct recall events in this period for this source."
          }
        },
        "type": "object",
        "required": [
          "period",
          "source",
          "event_count"
        ],
        "title": "PeriodCount",
        "description": "One (period, source) recall count — backs ``/stats/recalls-by-period`` (month/week/year)."
      },
...
      "StatsOverview": {
        "properties": {
          "total_recalls": {
            "type": "integer",
            "title": "Total Recalls",
            "description": "Count of recall events (rows in mart_recall_summary)."
          },
          "distinct_firms": {
            "type": "integer",
            "title": "Distinct Firms",
            "description": "Count of canonical firms (rows in mart_firm_profile)."
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources",
            "description": "The feeds covered: CPSC, FDA, USDA, NHTSA, USCG."
          },
          "last_rebuilt_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Rebuilt At",
            "description": "gold_meta.rebuilt_at — when the gold marts were last rebuilt (UTC)."
          }
        },
        "type": "object",
        "required": [
          "total_recalls",
          "distinct_firms",
          "sources"
        ],
        "title": "StatsOverview",
        "description": "Headline KPIs for the landing page (API-computed, not a stored ``fct_*``)."
      },
      "StatsSource": {
        "type": "string",
        "enum": [
          "CPSC",
          "FDA",
          "USDA",
          "NHTSA",
          "USCG",
          "ALL"
        ],
        "title": "StatsSource",
        "description": "The ``fct_*`` source domain — the 5 agency feeds plus the synthesized ``ALL`` rollup.\n\nDistinct from the strict ``Source`` enum (which has no ``ALL``). Per-endpoint population varies:\nmonthly-trend is per-source only (no ``ALL``); by-country is FDA/USDA (+``ALL``); units is\nNHTSA/USCG/FDA/USDA (no ``ALL``, no CPSC). A source absent from a given fact returns ``[]``."
      },
      "StatusCount": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Agency feed, or 'ALL'."
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "'active' / 'inactive' / 'unknown'. CPSC/NHTSA carry no lifecycle -> 'unknown'."
          },
          "event_count": {
            "type": "integer",
            "title": "Event Count"
          }
        },
        "type": "object",
        "required": [
          "source",
          "status",
          "event_count"
        ],
        "title": "StatusCount",
        "description": "Active / inactive / unknown recall counts per source (+ the 'ALL' rollup)."
      },
...
    }
  }
}